+++ resources/views/pages/superadmin/pages/organiser.blade.php
@@ -2,6 +2,8 @@
 @section('css')
         <link rel="stylesheet" href="//cdnjs.cloudflare.com/ajax/libs/material-design-lite/1.1.0/material.min.css">
         <link rel="stylesheet" href="https://cdn.datatables.net/1.10.15/css/dataTables.material.min.css">
+        <script src="https://code.jquery.com/jquery-1.12.4.js"></script>
+        <script src="{{ URL::to('public/admin/assets/js/bootstrap-notify.js') }}"></script>
 @endsection
 @section('contents')
  <div class="content">
@@ -10,11 +12,7 @@
         <div class="card-header" data-background-color="orange">
             <h4 class="title">Admin</h4>
         </div>
-            @if(Session::has('message'))
-                <div class="col-md-12">
-                <p class="alert {{ Session::get('alert-class', 'alert-info') }}">{{ Session::get('message') }}</p>
-                </div>
-            @endif
+            
             <div style="padding-right: 15px;" class="text-right">
                 <a href="{{ URL::to('newAdmin') }}" class="btn btn-warning">Create New Admin</a>
             </div>
@@ -27,7 +25,9 @@
                         <th class="text-center">Email</th>
                         <th class="text-center">Mobile</th>
                         <th class="text-center">Action</th>
+                        @if (!auth()->user()->isAdmin())
                         <th class="text-center">Action</th>
+                        @endif
                     </tr>
                 </thead>
                 <tbody>
@@ -41,12 +41,13 @@
                             <i class="material-icons">edit</i>
                             <div class="ripple-container"></div></a>
                         </td>
+                        @if (!auth()->user()->isAdmin())
                         <td>
-                             <a href="deleteAdmin{{$posts->id}}" type="button" rel="tooltip" class="btn btn-danger btn-xs" data-original-title="" title="Delete">
+                             <a href="deleteAdmin{{$posts->id}}" type="button" rel="tooltip" class="btn btn-danger btn-xs" data-original-title="" title="Delete" onclick="return ConfirmDelete()">
                             <i class="material-icons">delete</i>
                             <div class="ripple-container"></div></a>
                         </td>
-                        
+                        @endif
                     </tr>
                     @endforeach
                 </tbody>
@@ -57,6 +58,23 @@
     </div>
                 
 </div>
+@if(Session::has('message'))
+    <script>
+        var msg = "{{ Session::get('message') }}";
+            type = ['','info','success','warning','danger'];
+                $.notify({
+                icon: "notifications",
+                message: msg
+                },{
+                    type: 'success',
+                    timer: 1000,
+                    placement: {
+                        from: 'bottom',
+                        align: 'right'
+                    }
+                });
+    </script>
+@endif
 @endsection
 
 @section('javascript')
@@ -72,5 +90,20 @@
             }]
             });
             });
+        </script>
+        <script>
+        function ConfirmDelete()
+        {
+            var x = confirm("Are you sure you want to delete Category?");
+            if (x)
+            return true;
+            else
+            return false;
+        }
+        </script>
+        <script type="text/javascript">
+          setTimeout(function() {
+          $('#successMessage').fadeOut('slow');
+          }, 2000);
         </script>
 @endsection
