+++ storage/framework/views/27eb46f6d7d6c1ea6d1441b064d0e58813f46779.php
@@ -1,6 +1,9 @@
 <?php $__env->startSection('css'); ?>
         <link rel="stylesheet" href="//cdnjs.cloudflare.com/ajax/libs/material-design-lite/1.1.0/material.min.css">
         <link rel="stylesheet" href="https://cdn.datatables.net/1.10.15/css/dataTables.material.min.css">
+         <script src="https://code.jquery.com/jquery-1.12.4.js"></script>
+        <script src="<?php echo e(URL::to('public/admin/assets/js/bootstrap-notify.js')); ?>"></script>
+       
 <?php $__env->stopSection(); ?>
 <?php $__env->startSection('contents'); ?>
  <div class="content">
@@ -9,11 +12,7 @@
         <div class="card-header" data-background-color="orange">
             <h4 class="title">Page Management</h4>
         </div>
-            <?php if(Session::has('message')): ?>
-                <div class="col-md-12" id="successMessage">
-                <p class="alert <?php echo e(Session::get('alert-class', 'alert-info')); ?>"><?php echo e(Session::get('message')); ?></p>
-                </div>
-            <?php endif; ?>
+            
             <div style="padding-right: 15px;" class="text-right">
                 <a href="<?php echo e(URL::to('newPage')); ?>" class="btn btn-info">Create New Page</a>
             </div>
@@ -40,12 +39,12 @@
                         <td>
                              <a href="editPage<?php echo e($posts->id); ?>" type="button" rel="tooltip" class="btn btn-success btn-xs" data-original-title="" title="Edit">
                             <i class="material-icons">edit</i>
-                            <div class="ripple-container"></div></a>
+                            
                         </td>
                         <td>
-                             <a href="deletePage<?php echo e($posts->id); ?>" type="button" rel="tooltip" class="btn btn-danger btn-xs" data-original-title="" title="Delete">
+                             <a href="deletePage<?php echo e($posts->id); ?>" type="button" rel="tooltip" class="btn btn-danger btn-xs" data-original-title="" title="Delete" onclick="return ConfirmDelete()">
                             <i class="material-icons">delete</i>
-                            <div class="ripple-container"></div></a>
+                            
                         </td>
                         
                     </tr>
@@ -58,6 +57,23 @@
     </div>
                 
 </div>
+<?php if(Session::has('message')): ?>
+    <script>
+        var msg = "<?php echo e(Session::get('message')); ?>";
+            type = ['','info','success','warning','danger'];
+                $.notify({
+                icon: "notifications",
+                message: msg
+                },{
+                    type: 'success',
+                    timer: 1000,
+                    placement: {
+                        from: 'bottom',
+                        align: 'right'
+                    }
+                });
+    </script>
+<?php endif; ?>
 <?php $__env->stopSection(); ?>
 
 <?php $__env->startSection('javascript'); ?>
@@ -73,6 +89,16 @@
             }]
             });
             });
+        </script>
+        <script>
+         function ConfirmDelete()
+        {
+            var x = confirm("Are you sure you want to delete Category?");
+            if (x)
+            return true;
+            else
+            return false;
+        }
         </script>
         <script type="text/javascript">
           setTimeout(function() {
